requirements.txt
requests
pytests
argparse